Task 0 and 1 corrections

+++ 0x1A-hash_tables/0-main.c
@@ -9,7 +9,7 @@ int main(void)
 {
     hash_table_t *ht;
 
-    ht = (int)hash_table_create(1024);
+    ht = hash_table_create(1024);
     printf("%p\n", (void *)ht);
     return (EXIT_SUCCESS);
 }
